requirements.txt
asgiref==3.5.2
comtypes==1.1.10
dj-config-url==0.1.1
Django==4.0.6
django-cors-headers==3.14.0
django-debug-toolbar==3.6.0
django-environ==0.9.0
django-rest-registration==0.8.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.2
evdev==1.6.1
gunicorn==20.1.0
protobuf==4.22.1
psycopg2==2.9.7
PyJWT==1.7.1
pynput==1.7.3
python-dateutil==2.8.1
python-xlib==0.33
pytz==2021.1
selenium==3.141.0
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.6